pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[project]
name = "agent-ia"
version = "0.1.0"
description = "Application d'agent IA pour l'immobilier"
authors = [{name = "Herry Fabien", email = "fabien_herry@yahoo.fr"},{"name"="Arnaud Rambourd"}]
license = {text = "Undefined"}
dependencies = [
    "python-dotenv",
    "langchain",
    "langchain-mistralai",
    "psycopg[binary]",
    "fastmcp>=3.2.0",
    "fastapi",
    "uvicorn",
    "requests",
]
